refactor: move config related files to the api package
